+++ api/src/server.rs
@@ -28,7 +28,8 @@ use ceramic_api_server::{
     InterestsSortKeySortValuePostResponse, LivenessGetResponse, VersionPostResponse,
 };
 use ceramic_api_server::{
-    Api, ExperimentalEventsSepSepValueGetResponse, FeedEventsGetResponse, InterestsPostResponse,
+    Api, ExperimentalEventsSepSepValueGetResponse, ExperimentalInterestsGetResponse,
+    FeedEventsGetResponse, InterestsPostResponse,
 };
 use ceramic_core::{Cid, EventId, Interest, Network, PeerId, StreamId};
 use futures::TryFutureExt;
@@ -129,8 +130,9 @@ impl TryFrom<models::Interest> for ValidatedInterest {
     }
 }
 
+/// Trait for accessing persistent storage of Interests
 #[async_trait]
-pub trait AccessInterestStore: Send + Sync {
+pub trait InterestStore: Send + Sync {
     /// Returns true if the key was newly inserted, false if it already existed.
     async fn insert(&self, key: Interest) -> Result<bool>;
     async fn range(
@@ -143,7 +145,7 @@ pub trait AccessInterestStore: Send + Sync {
 }
 
 #[async_trait]
-impl<S: AccessInterestStore> AccessInterestStore for Arc<S> {
+impl<S: InterestStore> InterestStore for Arc<S> {
     async fn insert(&self, key: Interest) -> Result<bool> {
         self.as_ref().insert(key).await
     }
@@ -159,8 +161,9 @@ impl<S: AccessInterestStore> AccessInterestStore for Arc<S> {
     }
 }
 
+/// Trait for accessing persistent storage of Events
 #[async_trait]
-pub trait AccessModelStore: Send + Sync {
+pub trait EventStore: Send + Sync {
     /// Returns (new_key, new_value) where true if was newly inserted, false if it already existed.
     async fn insert_many(&self, items: &[(EventId, Vec<u8>)]) -> Result<Vec<bool>>;
     async fn range_with_values(
@@ -192,7 +195,7 @@ pub trait AccessModelStore: Send + Sync {
 }
 
 #[async_trait::async_trait]
-impl<S: AccessModelStore> AccessModelStore for Arc<S> {
+impl<S: EventStore> EventStore for Arc<S> {
     async fn insert_many(&self, items: &[(EventId, Vec<u8>)]) -> Result<Vec<bool>> {
         self.as_ref().insert_many(items).await
     }
@@ -253,8 +256,8 @@ pub struct Server<C, I, M> {
 
 impl<C, I, M> Server<C, I, M>
 where
-    I: AccessInterestStore,
-    M: AccessModelStore + 'static,
+    I: InterestStore,
+    M: EventStore + 'static,
 {
     pub fn new(peer_id: PeerId, network: Network, interest: I, model: Arc<M>) -> Self {
         let (tx, event_rx) = tokio::sync::mpsc::channel::<EventInsert>(1024);
@@ -374,6 +377,30 @@ where
             resume_token: new_hw.to_string(),
             events,
         }))
+    }
+
+    pub async fn get_interests(&self) -> Result<ExperimentalInterestsGetResponse, ErrorResponse> {
+        let interests = self
+            .interest
+            .range(
+                &Interest::min_value(),
+                &Interest::max_value(),
+                0,
+                usize::MAX,
+            )
+            .await
+            .map_err(|e| ErrorResponse::new(format!("failed to get interests: {e}")))?;
+
+        Ok(ExperimentalInterestsGetResponse::Success(
+            models::InterestsGet {
+                interests: interests
+                    .into_iter()
+                    .map(|i| models::InterestsGetInterestsInner {
+                        data: i.to_string(),
+                    })
+                    .collect(),
+            },
+        ))
     }
 
     pub async fn get_events_sort_key_sort_value(
@@ -609,8 +636,8 @@ pub(crate) fn decode_multibase_data(value: &str) -> Result<Vec<u8>, BadRequestRe
 impl<C, I, M> Api<C> for Server<C, I, M>
 where
     C: Send + Sync,
-    I: AccessInterestStore + Sync,
-    M: AccessModelStore + Sync + 'static,
+    I: InterestStore + Sync,
+    M: EventStore + Sync + 'static,
 {
     #[instrument(skip(self, _context), ret(level = Level::DEBUG), err(level = Level::ERROR))]
     async fn liveness_get(
@@ -669,6 +696,15 @@ where
             .or_else(|err| Ok(FeedEventsGetResponse::InternalServerError(err)))
     }
 
+    async fn experimental_interests_get(
+        &self,
+        _context: &C,
+    ) -> Result<ExperimentalInterestsGetResponse, ApiError> {
+        self.get_interests()
+            .await
+            .or_else(|err| Ok(ExperimentalInterestsGetResponse::InternalServerError(err)))
+    }
+
     #[instrument(skip(self, _context), ret(level = Level::DEBUG), err(level = Level::ERROR))]
     async fn experimental_events_sep_sep_value_get(
         &self,
@@ -710,14 +746,14 @@ where
     async fn interests_sort_key_sort_value_post(
         &self,
         sep_key: String,
-        seplue: String,
+        sep_value: String,
         controller: Option<String>,
         stream_id: Option<String>,
         _context: &C,
     ) -> Result<InterestsSortKeySortValuePostResponse, ApiError> {
         let interest = models::Interest {
             sep: sep_key,
-            sep_value: seplue,
+            sep_value,
             controller,
             stream_id,
         };